overloaded count functions to accept a single argument of one of three possible types: str, NormalizedReference, or List[NormalizedReference]
